we  add check() to the inital (nonsensical test) with the corrected Sensor, the situation is now the following: the test is randomly green and red!
and Alarm's code for SOLID violations, hidden dependencies.
All the logic of the Alarm class is in the check method.

Alarm's main problem is that it depends on a concrete class, the sensor that measures pressure values.

This is a clear violation of the Dependency Inversion principle (DIP) which states that:
Abstractions should not depend on details.

Details should depend on abstractions.
To make things worse this dependency is hidden inside Alarm's check method (see J. B. Rainsberger's The Pain of Implicit Dependencies post).

The violation of the DIP and the implicit dependency makes also impossible to fulfill the Open Closed Principle (OCP).

Fulfilling those two principles will be one of the main goals of this refactoring.
ref: http://garajeando.blogspot.de/2015/08/tire-pressure-monitoring-system.html

236deb3 Eduard Ruffert At Home <ruffert@gmail.com> on 16/03/2017 at 20:49